refactor(app): move AppContent to an exclusive file
